Rangzen performs exchanges (in the clear) with nearby devices!
Tested only with one other device, some bugs exist.

---

Removed tabs from MessageStore. Not sure how they got there.

Shorten long lines.

Synchronize over mSeeking in WifiDirectSpeaker.

Make MessageStore.checkPriority package access and test it.

Added aliases to all tests in .buckconfig

Added more tests for priority calculation.

Exchange priority calculation uses stored priority if higher.

Fixed tests and mis-ordered instance vars in Opener.

Wait between RECEIVING Exchanges as well as starting them.

This may help against a bug I encountered where the devices got into a
loop of failing to exchange once a second, disregarding that they should
have timed out. I don't think it really fixes the bug.

Commented code removal and log messages.

Don't call .close() on socket if socket is null.

No longer divide by 0 in new priority calculations - use epsilon.

Fixed null pointer in Exchange when friends received are empty.

Fixed argument order for newPriority call in RangzenService.

Convert floats to doubles for messages, and implement a real priority fn.

Add some aliases for tests to .buckconfig.

Store time of exchange in StorageBase, only have exchanges every 10 sec.

Added putLong/getLong methods to StorageBase.

Exchange is Runnable instead of AsyncTask.

AsyncTask just isn't a fit for what we were doing - AsyncTasks should be
run from the UI.

Choose random peer for exchange instead of 0th. Better comments.

Cleaned up some syntax with shorter name for RangzenMessage.

Modified protobuf building to work with protobuf imports.

And changed protobufs to use separate RangzenMessage class.

Now must add each new protobuf file to protos.include. This is
inconvenient but is the easiest way to make the protobuf compiler
notice all the files that depend on one another.

Whitespace etc.

WireTest was broken due to double constant instaed of float. Fixed.

It's alive! Messages are transmitted and put into the feed.

Added updatePriority() and contains() to MessageStore.

Added getPriority(String) method to messageStore.

Gets the priority for a message, or NOT_FOUND if the message isn't in
the store.

Have cleartext exchanges with found peers. Logs show messages passing!

R=jesus@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/6180001

+++ ui/Rangzen/src/org/denovogroup/rangzen/FeedListAdapter.java
@@ -121,7 +121,7 @@ public class FeedListAdapter extends BaseAdapter {
             mViewHolder = (ViewHolder) convertView.getTag();
         }
         mViewHolder.mHashtagView.setText(message.getMessage());
-        mViewHolder.mUpvoteView.setText(Float.toString(message.getPriority()));
+        mViewHolder.mUpvoteView.setText(Double.toString(message.getPriority()));
 
         return convertView;
     }
